requirements.txt
flask
flask-cors
ultralytics
numpy
pillow
gunicorn
